requirements.txt
numpy
torch
torchvision
python-mnist
pillow
opencv-python
matplotlib
seaborn
tensorboardX
antspyx
SimpleITK